backend/requirements.txt
Flask==2.3.3
python-oracledb
PyJWT==2.8.0
python-dotenv==1.0.0
Flask-CORS==4.0.0
gunicorn==21.2.0